a ton of refactoring and improvements
