+++ FargoSoulsPlayer.cs
@@ -414,7 +414,6 @@ namespace FargowiltasSouls
         public byte lightningRodTimer;
         public bool ReverseManaFlow;
         public bool CurseoftheMoon;
-        public bool FadingSoul;
         public bool OceanicMaul;
         public int MaxLifeReduction;
         public int CurrentLifeReduction;
@@ -961,7 +960,6 @@ namespace FargowiltasSouls
             Jammed = false;
             ReverseManaFlow = false;
             CurseoftheMoon = false;
-            FadingSoul = false;
             OceanicMaul = false;
             DeathMarked = false;
             Hypothermia = false;
@@ -1903,10 +1901,6 @@ namespace FargowiltasSouls
 
             if (CurseoftheMoon)
                 DamageOverTime(20);
-            if (FadingSoul)
-            {
-                DamageOverTime(200);
-            }
 
             if (Oiled && Player.lifeRegen < 0)
             {
@@ -2122,29 +2116,13 @@ namespace FargowiltasSouls
                 }
             }
 
-            if (FadingSoul)
-            {
-                if (Main.rand.NextBool(4) && drawInfo.shadow == 0f)
-                {
-                    int dust2 = Dust.NewDust(drawInfo.Position - new Vector2(2f, 2f), Player.width + 4, Player.height + 4, DustID.Water_Hallowed, Player.velocity.X * 0.4f, Player.velocity.Y * 0.4f, 100, default(Color), 3f);
-                    Main.dust[dust2].noGravity = true;
-                    Main.dust[dust2].velocity *= 1.8f;
-                    Main.dust[dust2].velocity.Y -= 0.5f;
-                    drawInfo.DustCache.Add(dust2);
-                }
-                r *= 0.1f;
-                g *= 0.2f;
-                b *= 0.7f;
-                fullBright = true;
-            }
-
             if (DeathMarked)
             {
                 if (Main.rand.NextBool() && drawInfo.shadow == 0f)
                 {
                     int dust = Dust.NewDust(Player.position - new Vector2(2f, 2f), Player.width, Player.height, 109, Player.velocity.X * 0.4f, Player.velocity.Y * 0.4f, 0, default(Color), 1.5f);
                     Main.dust[dust].velocity.Y--;
-                    if (Main.rand.Next(3) != 0)
+                    if (!Main.rand.NextBool(3))
                     {
                         Main.dust[dust].noGravity = true;
                         Main.dust[dust].scale += 0.5f;
