+++ Localization/en-US.hjson
@@ -129,6 +129,7 @@ Mods: {
 			CosmosChampionKilledAll: Eridanus, Champion of Cosmos has slain every foe.
 
 			TrojanSquirrelDespawn: Trojan Squirrel returns to its squirrel clan...
+			LifeChallengerDespawn: Lifelight fades away...
 		}
 
 		Message: {
