+++ Localization/pt-BR.hjson
@@ -3269,6 +3269,7 @@ Mods: {
 			CosmosChampionKilledAll: Eridanus, Campeão do Cosmos matou todos os inimigos.
 			
 			TrojanSquirrelDespawn: O Campeão de Madeira retorna a seu clã esquilo.
+			LifeChallengerDespawn: Lifelight desvanece...
 		}
 
 		Message: {
