+++ Localization/zh-Hans.hjson
@@ -118,6 +118,7 @@ Mods: {
 			CosmosChampionKilledAll: 宇宙英灵击败了每一个敌人。
 
 			TrojanSquirrelDespawn: 特洛伊松鼠回到了它的松鼠聚落中……
+			LifeChallengerDespawn: 生命之光消失在辉光中……
 		}
 
 		NPCName:{
