+++ Projectiles/Challengers/LifeTpTelegraph.png
